Add GTM script and remove old GA code
