+++ cla_public/config/common.py
@@ -75,11 +75,6 @@ ZENDESK_API_USERNAME = os.environ.get("ZENDESK_API_USERNAME")
 ZENDESK_API_TOKEN = os.environ.get("ZENDESK_API_TOKEN")
 ZENDESK_DEFAULT_REQUESTER = 649762516  # anonymous feedback <noreply@ministryofjustice.zendesk.com>
 
-GDS_GA_ID = os.environ.get("GDS_GA_ID")
-MOJ_GTM_ID = os.environ.get("MOJ_GTM_ID")
-MOJ_GTM_AUTH = os.environ.get("MOJ_GTM_AUTH")
-MOJ_GTM_PREVIEW = os.environ.get("MOJ_GTM_PREVIEW")
-
 CACHE_TYPE = "simple"
 
 EXTENSIONS = []
