+++ cla_public/static-src/javascripts/modules/confirmation-print.js
@@ -17,9 +17,6 @@
 
       this.printButton.on('click', function() {
         window.print();
-        if(window.ga) {
-          window.ga('send', 'event', 'confirmation', 'printed');
-        }
       });
     },
