+++ cla_public/static-templates/errors/50x.html
@@ -7,6 +7,9 @@
       {% trans %}Service unavailable - Civil Legal Advice - GOV.UK{% endtrans %}
     </title>
     <script nonce="{{ csp_nonce() }}" type="text/javascript">
+        (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src='https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);})(window,document,'script','dataLayer','GTM-M9QXCZWK');
+    </script>
+    <script nonce="{{ csp_nonce() }}" type="text/javascript">
       (function(){if(navigator.userAgent.match(/IEMobile\/10\.0/)){var d=document,c="appendChild",a=d.createElement("style");a[c](d.createTextNode("@-ms-viewport{width:auto!important}"));d.getElementsByTagName("head")[0][c](a);}})();
     </script>
     <!--[if gt IE 8]><!--><link href="/moj-static/stylesheets/govuk-template.css" media="screen" rel="stylesheet" type="text/css" /><!--<![endif]-->
@@ -64,15 +67,6 @@
     <!--[if IE 8]>
         <link rel="stylesheet" href="/static/stylesheets/ie/ie8.min.css">
     <![endif]-->
-    <script nonce="{{ csp_nonce() }}">
-      (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-      (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-      m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-      })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-      ga('create', 'UA-37377084-13', 'auto');
-      ga('set', 'anonymizeIp', true);
-      ga('send', 'pageview');
-    </script>
   </head>
   <body class="govuk-template__body">
       <script nonce="{{ csp_nonce() }}" type="text/javascript">
