--- cla_public/templates/_ga-outcome-pageview.html
@@ -1,3 +1,0 @@
-{% if session.stored.outcome -%}
-  <script nonce="{{ csp_nonce() }}">ga('send', 'pageview', '/vpv/outcome/{{ session.stored.outcome }}');</script>
-{%- endif %}
