+++ cla_public/templates/checker/result/eligible-f2f.html
@@ -50,7 +50,6 @@
   {% block javascripts %}
     {{ super() }}
     <script nonce="{{ csp_nonce() }}" type="text/javascript" src="//maps.googleapis.com/maps/api/js?sensor=true&key={{config.GOOGLE_MAPS_API_KEY}}"></script>
-    {% include '_ga-outcome-pageview.html' %}
   {% endblock %}
 
 {% endif %}
