+++ cla_public/templates/checker/result/face-to-face.html
@@ -30,12 +30,6 @@
   {% block javascripts %}
     {{ super() }}
     <script nonce="{{ csp_nonce() }}" type="text/javascript" src="//maps.googleapis.com/maps/api/js?sensor=true&key={{config.GOOGLE_MAPS_API_KEY}}"></script>
-    {% include '_ga-outcome-pageview.html' %}
   {% endblock %}
 
-  {% block ga_pageview -%}
-    ga('moj.set', 'title', 'Results for [postcode] | Seek legal advice');
-    ga('govuk_shared.set', 'title', 'Results for [postcode] | Seek legal advice');
-    {{ super() }}
-  {%- endblock %}
 {% endif %}
