+++ cla_public/templates/checker/result/ineligible.html
@@ -53,8 +53,7 @@
 
   {% if category == 'family' %}
     <p class="govuk-body">
-      {% set ga_event = 'event:%s/%s/%s' % ('External Link Clicked', 'click', 'Family Mediation Council') %}
-      {% trans family_mediation_link=Element.link_same_window('http://www.familymediationcouncil.org.uk/', _('Family Mediation Council'), True, **{'data-ga': ga_event}) %}If you want to make an application to court about a family matter you need to first of all see if family mediation will help. Use the{% endtrans %} <a class="govuk-link" href="http://www.familymediationcouncil.org.uk">{% trans %}Family Mediation Council{% endtrans %}</a> {% trans %}directory to find a mediator and make an appointment for a Mediation Information and Assessment Meeting (MIAM).{% endtrans %}
+      {% trans family_mediation_link=Element.link_same_window('http://www.familymediationcouncil.org.uk/', _('Family Mediation Council'), True) %}If you want to make an application to court about a family matter you need to first of all see if family mediation will help. Use the{% endtrans %} <a class="govuk-link" href="http://www.familymediationcouncil.org.uk">{% trans %}Family Mediation Council{% endtrans %}</a> {% trans %}directory to find a mediator and make an appointment for a Mediation Information and Assessment Meeting (MIAM).{% endtrans %}
     </p>
   {% endif %}
 
@@ -85,17 +84,9 @@
   {% endif %}
 
   <p class="govuk-body">
-    {% set ga_event = 'event:%s/%s/%s' % ('External Link Clicked', 'click', 'Find a solicitor') %}
-    {{ Element.link_same_window('https://www.gov.uk/find-a-legal-adviser', _('Find a solicitor'), **{'class': 'govuk-button', 'data-ga': ga_event}) }}
+    {{ Element.link_same_window('https://www.gov.uk/find-a-legal-adviser', _('Find a solicitor'), **{'class': 'govuk-button'}) }}
   </p>
 
   {{ HelpOrganisations.org_list_container(organisations, category_name, truncate=5) }}
   {% include "checker/time-out-warning.html" %}
 {% endblock %}
-
-{% block javascripts %}
-  {{ super() }}
-  {% include '_ga-outcome-pageview.html' %}
-{% endblock %}
-
-
