+++ cla_public/templates/errors/404.html
@@ -7,18 +7,9 @@
   <p class="govuk-body">{{ _('Please return to the <a class="govuk-link" href="/">Can I get legal aid?</a> page and try again.') }}</p>
 {% endblock %}
 
-{% block javascripts %}
-  <!-- Google analytics -->
+{% block head %}
   <script nonce="{{ csp_nonce() }}">
-    (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-      (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-      m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-    })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-    function stripPII(input) {
-      return input.replace(/postcode=[^&]+/gi, 'postcode=[postcode]');
-    }
-    ga('create', 'UA-37377084-21', 'auto');
-    ga('send', 'pageview');
-    ga('send', 'event', 'errors', '404', window.location.pathname + stripPII(window.location.search), 1);
+    GTM_ANON_ID = '{{ GTM_ANON_ID }}';
   </script>
+  <script type="text/javascript" nonce="{{ csp_nonce() }}" src="{{ asset(filename='javascripts/google-tag-manager.js') }}"></script>
 {% endblock %}
