+++ cla_public/templates/index.html
@@ -18,7 +18,7 @@
 
   <div role="note" aria-label="Information" class="govuk-inset-text">
     <p>
-      {{ _('This guide is also available <a class="govuk-link" lang="cy" href="/locale/cy_GB" data-ga="event:locale/switch/welsh">in Welsh (Cymraeg)</a>') }}.
+      {{ _('This guide is also available <a class="govuk-link" lang="cy" href="/locale/cy_GB">in Welsh (Cymraeg)</a>') }}.
     </p>
   </div>
