+++ cla_public/templates/ineligible.html
@@ -7,6 +7,5 @@
 
 {% block javascripts %}
   {{ super() }}
-  {% include '_ga-outcome-pageview.html' %}
   {{ CAIT.script(info_tools, cait_variant, cait_journey, cait_js) }}
 {% endblock %}
