+++ cla_public/templates/interstitial.html
@@ -80,8 +80,7 @@
     {{ _('If you qualify, in some cases you may need to pay a contribution towards your legal aid.') }}
   </p>
   <p class="govuk-body">
-    {% set ga_event = 'event:%s/%s/%s' % ('External Link Clicked', 'click', 'Find a legal adviser') %}
-    <a class="govuk-link" href="{{ url_for('.laalaa') }}?category={{ category }}" data-ga="{{ ga_event }}">{{ _('Find a legal adviser') }}</a>
+    <a class="govuk-link" href="{{ url_for('.laalaa') }}?category={{ category }}">{{ _('Find a legal adviser') }}</a>
   </p>
 
   {% if organisations %}
