+++ cla_public/templates/moj_template_base.jinja
@@ -152,11 +152,9 @@
         <ul class="govuk-footer__inline-list">
           <li class="govuk-footer__inline-list-item">
             {% if request.cookies.get('locale') == 'cy_GB' %}
-              <a class="govuk-footer__link" lang="en" href="{{ url_for('base.set_locale', locale='en_GB') }}"
-                 data-ga="event:locale/switch/english">English</a>
+              <a class="govuk-footer__link" lang="en" href="{{ url_for('base.set_locale', locale='en_GB') }}">English</a>
             {% else %}
-              <a class="govuk-footer__link" lang="cy" href="{{ url_for('base.set_locale', locale='cy_GB') }}"
-                 data-ga="event:locale/switch/welsh">Cymraeg</a>
+              <a class="govuk-footer__link" lang="cy" href="{{ url_for('base.set_locale', locale='cy_GB') }}">Cymraeg</a>
             {% endif %}
           </li>
           {% block footer_support_links %}{% endblock %}
