+++ cla_public/templates/scope/diagnosis.html
@@ -24,7 +24,7 @@
     <ul class="cla-scope-options-list">
       {% for choice in choices %}
         <li class="cla-scope-options-list-item">
-          <a class="cla-scope-options-list-item-link" title="{{ choice.label|safe|striptags }}" href="{{ choice.url }}"{% if not nodes|length %} data-ga="event:category-of-law/{{ choice.label|striptags }}"{% endif %} role="button">
+          <a class="cla-scope-options-list-item-link" title="{{ choice.label|safe|striptags }}" href="{{ choice.url }}" role="button">
             <h2 class="govuk-heading-m govuk-!-margin-2">{{ choice.label|safe|striptags }}</h2>
             {% if choice.help_text %}
               <p class="govuk-body govuk-!-margin-2">{{ choice.help_text|safe|striptags }}</p>
