+++ cla_public/templates/scope/mediation.html
@@ -31,8 +31,3 @@
   </p>
   {% include "checker/time-out-warning.html" %}
 {% endblock %}
-
-{% block javascripts %}
-  {{ super() }}
-  {% include '_ga-outcome-pageview.html' %}
-{% endblock %}
