+++ helm_deploy/cla-public/values-dev.yaml
@@ -15,8 +15,6 @@ environment: development
 
 # Lists don't deep merge, so this list of envVars overrides anything defined in an earlier values file
 envVars:
-  GDS_GA_ID:
-    value: UA-145652997-1
   BACKEND_BASE_URI:
     value: "http://cla-backend-app.laa-cla-backend-staging.svc.cluster.local"
   LAALAA_API_HOST:
