+++ tasks/_paths.js
@@ -8,7 +8,9 @@ var paths = {
   scripts: [
     'cla_public/static-src/javascripts/**/*',
     '!cla_public/static-src/javascripts/templates.js',
+    '!cla_public/static-src/javascripts/headscripts/*',
   ],
+  head_scripts: 'cla_public/static-src/javascripts/headscripts/*',
   vendor_scripts: ['cla_public/static-src/vendor/*'],
   images: 'cla_public/static-src/images/**/*',
   templates: 'cla_public/templates/**/*',
